phresh-tutorial/backend/requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
pydantic==1.8.2
email-validator==1.1.3
databases[postgresql]==0.4.3
SQLAlchemy==1.3
alembic==1.6.5
psycopg2-binary==2.9.1
pyjwt==2.1.0
passlib[bcrypt]~=1.7.4
pytest==6.2.1
pytest-asyncio==0.14.0
httpx==0.16.1
asgi-lifespan==1.0.1
debugpy==1.3.0